Modulo 10 exercicio, alteração 3

+++ BACK-END/modulo-10/tarefa_controledefluxo/Main.java
@@ -3,22 +3,22 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
 
-        int a, b, c, d, notaMedia;
+        float a, b, c, d, notaMedia;
 
         Scanner read = new Scanner(System.in);
 
 
         System.out.println("Insira sua primeira nota:");
-        a = read.nextInt();
+        a = read.nextFloat();
 
         System.out.println("Insira sua segunda nota:");
-        b = read.nextInt();
+        b = read.nextFloat();
 
         System.out.println("Insira sua terceira nota:");
-        c = read.nextInt();
+        c = read.nextFloat();
 
         System.out.println("Insira sua quarta nota:");
-        d = read.nextInt();
+        d = read.nextFloat();
 
         notaMedia = (a + b + c + d) / 4;
